Issue 18 Make it all work on Java 11 (#19)
* Compiles, test suite and REST service with Swagger UI run with java.version=11 on JDK 11.

* Uses manual DODI instead of AspectJ-enabled DODI.

* Maven build and site generation work on Java 11.
./mvnw clean test package succeeds if all runs on JDK 11.
./mvnw site succeeds if all runs on JDK 11.

+++ src/test/java/de/beuth/knabe/spring_ddd_bank/infrastructure/AccountJpaRepositoryTest.java
@@ -52,8 +52,8 @@ public class AccountJpaRepositoryTest {
 		assertThat(annasAccountId, greaterThan(jacksAccountId));
 		
 		//The accountNo is always the same as the accountId:
-		assertEquals(annasAccountId, annasAccount.accountNo().toLong());
-		assertEquals(jacksAccountId, jacksAccount.accountNo().toLong());
+		assertEquals(annasAccountId.longValue(), annasAccount.accountNo().toLong());
+		assertEquals(jacksAccountId.longValue(), jacksAccount.accountNo().toLong());
     }
 
 }
